+++ src/main/java/randommcsomethin/fallingleaves/FallingLeavesClient.java
@@ -3,18 +3,18 @@ package randommcsomethin.fallingleaves;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import randommcsomethin.fallingleaves.init.Config;
 import randommcsomethin.fallingleaves.init.Leaves;
-import randommcsomethin.fallingleaves.util.LogUtil;
 
 @Environment(EnvType.CLIENT)
 public class FallingLeavesClient implements ClientModInitializer {
 
-    // The mod's unique identifier, used when interacting with the registry.
-    public static String MOD_ID = "fallingleaves";
+    /** The mod's unique identifier, used to avoid mod conflicts in both the registry and config files */
+    public static final String MOD_ID = "fallingleaves";
 
-    // The logger object, a facade/proxy for Log4j. Singleton here for ease of use.
-    public static LogUtil LOGGER = new LogUtil(MOD_ID);
+    public static final Logger LOGGER = LogManager.getLogger(MOD_ID);
 
     @Override
     public void onInitializeClient() {