+++ src/main/java/randommcsomethin/fallingleaves/util/RegistryUtil.java
@@ -1,13 +1,17 @@
 package randommcsomethin.fallingleaves.util;
 
 import net.fabricmc.fabric.api.particle.v1.FabricParticleTypes;
+import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.particle.DefaultParticleType;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.registry.Registry;
+import org.jetbrains.annotations.Nullable;
 import randommcsomethin.fallingleaves.FallingLeavesClient;
 
-import static randommcsomethin.fallingleaves.util.LogUtil.LOGGER;
+import java.util.Optional;
+
+import static randommcsomethin.fallingleaves.FallingLeavesClient.LOGGER;
 
 /**
  * TODO - Explore ParticleTypes in the registry. Should we be using it to
@@ -31,6 +35,12 @@ public class RegistryUtil {
 
     public static String getBlockId(BlockState blockState) {
         return Registry.BLOCK.getId(blockState.getBlock()).toString();
+    }
+
+    @Nullable
+    public static Block getBlock(String blockId) {
+        Optional<Block> maybeBlock = Registry.BLOCK.getOrEmpty(new Identifier(blockId));
+        return maybeBlock.orElse(null);
     }
 
 }